moved fetching statistics to different component
