+++ src/js/components/Peak/PeakDescription.tsx
@@ -1,21 +1,11 @@
 import styles from './PeakDescription.module.css';
 import React from 'react';
-import { Peak, User } from '../../models/interfaces';
+import { Peak } from '../../models/interfaces';
 import { SubmitButton } from '../common';
 import { completeThePeak } from '../../API/peaks/methods';
 import PeakStatistics from './PeakStatistics';
 
-export type PeakDescriptionProps = {
-  peak: Peak;
-  firstConqueror: User | undefined;
-  numberOfPeakConquerors: number | undefined;
-};
-
-const PeakDescription = ({
-  peak,
-  firstConqueror,
-  numberOfPeakConquerors,
-}: PeakDescriptionProps) => {
+const PeakDescription = ({ peak }: { peak: Peak }) => {
   // TODO add form to reach peak
   const reachPeak = async (time: number) => {
     const peakCompletionResponse = await completeThePeak(peak.id, time);
@@ -29,10 +19,7 @@ const PeakDescription = ({
       <p>Wysokość: {peak.height} m n.p.m.</p>
       <p>Województwo: {peak.region}</p>
       <p>Pasmo górskie: {peak.mountainRange}</p>
-      <PeakStatistics
-        firstConqueror={firstConqueror}
-        numberOfPeakConquerors={numberOfPeakConquerors}
-      />
+      <PeakStatistics peakId={peak.id} />
       <div className={styles.buttonBox}>
         <SubmitButton
           onClick={() => reachPeak(3000)}
